+++ Samples~/MainSample/Scripts/UI/EffectTypeDropdown.cs
@@ -1,0 +1,19 @@
+﻿using UnityEngine;
+
+namespace DualSenseSample.UI
+{
+    public class EffectTypeDropdown : MonoBehaviour
+    {
+        public GameObject[] options;
+
+        public void SetOption(int index)
+        {
+            foreach (var option in options)
+            {
+                option.SetActive(false);
+            }
+
+            options[index].SetActive(true);
+        }
+    }
+}