+++ UniSense/DualSenseGamepadHID.cs
@@ -21,7 +21,7 @@ namespace UniSense
         public ButtonControl leftTriggerButton { get; protected set; }
         public ButtonControl rightTriggerButton { get; protected set; }
         public ButtonControl playStationButton { get; protected set; }
-        
+
         public ButtonControl micMuteButton { get; protected set; }
 
 #if UNITY_EDITOR
@@ -30,6 +30,29 @@ namespace UniSense
             Initialize();
         }
 #endif
+
+        /// <summary>
+        /// Finds the first DualSense connected by the player or <c>null</c> if 
+        /// there is no one connected to the system.
+        /// </summary>
+        /// <returns>A DualSenseGamepadHID instance or <c>null</c>.</returns>
+        public static DualSenseGamepadHID FindFirst()
+        {
+            foreach (var gamepad in all)
+            {
+                var isDualSenseGamepad = gamepad is DualSenseGamepadHID;
+                if (isDualSenseGamepad) return gamepad as DualSenseGamepadHID;
+            }
+
+            return null;
+        }
+
+        /// <summary>
+        /// Finds the DualSense last used/connected by the player or <c>null</c> if 
+        /// there is no one connected to the system.
+        /// </summary>
+        /// <returns>A DualSenseGamepadHID instance or <c>null</c>.</returns>
+        public static DualSenseGamepadHID FindCurrent() => Gamepad.current as DualSenseGamepadHID;
 
         [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.BeforeSceneLoad)]
         static void Initialize()
@@ -85,6 +108,27 @@ namespace UniSense
             m_LowFrequencyMotorSpeed = null;
         }
 
+        public void ResetMotorSpeeds() => SetMotorSpeeds(0f, 0f);
+
+        public void ResetLightBarColor() => SetLightBarColor(Color.black);
+
+        public void ResetTriggersState()
+        {
+            var command = DualSenseHIDOutputReport.Create();
+            command.SetRightTriggerState(m_rightTriggerState.Value);
+            command.SetLeftTriggerState(m_leftTriggerState.Value);
+
+            ExecuteCommand(ref command);
+        }
+
+        public void Reset()
+        {
+            ResetHaptics();
+            ResetMotorSpeeds();
+            ResetLightBarColor();
+            ResetTriggersState();
+        }
+
         public override void ResumeHaptics()
         {
             if (!MotorHasValue && !LeftTriggerHasValue && !RightTriggerHasValue)
@@ -126,7 +170,7 @@ namespace UniSense
                 var lightBarColor = state.LightBarColor.Value;
                 command.SetLightBarColor(lightBarColor);
             }
-            
+
             if (state.Motor.HasValue)
             {
                 var motor = state.Motor.Value;
@@ -166,7 +210,7 @@ namespace UniSense
                 var playerLed = state.PlayerLed.Value;
                 command.SetPlayerLedState(playerLed);
             }
-                
+
             ExecuteCommand(ref command);
         }